Change all NPC speeds into pixel per 1/65 sec. To have more convenient way to apply original NPC speeds from legacy engine.

+++ Engine/scenes/level/lvl_npc.h
@@ -44,7 +44,7 @@ public:
     void setDirection(int dir);
     int  direction();
     int _direction;
-    float motionSpeed;
+    double motionSpeed;
     bool  is_scenery;
     bool  is_activity;
     bool  is_shared_animation;
